Refactored code, separated header files from source files. Included tinyxml as a third-party library in code. Configured CMakeLists.txt files for updated code structure.
