+++ include/config.h
@@ -2,8 +2,8 @@
 #define CONFIG_H
 
 #include "gl_const.h"
-#include "tinyxml/tinyxml.h"
-#include "tinyxml/tinystr.h"
+#include "tinyxml.h"
+#include "tinystr.h"
 
 #include <algorithm>
 #include <iostream>
