+++ include/openlist.h
@@ -2,8 +2,8 @@
 #define OPENLIST_H
 
 #include "node.h"
-#include "tinyxml/tinyxml.h"
-#include "tinyxml/tinystr.h"
+#include "tinyxml.h"
+#include "tinystr.h"
 
 #include <vector>
