+++ include/searchresult.h
