+++ main.cpp
@@ -1,4 +1,4 @@
-#include"mission.h"
+#include "include/mission.h"
 
 #include <iostream>
